+++ Kernel/Language/pt.pm
@@ -4785,6 +4785,8 @@ sub Data {
         'Can\'t read file!' => '',
         '<p>If you continue to install this package, the following issues may occur:</p><ul><li>Security problems</li><li>Stability problems</li><li>Performance problems</li></ul><p>Please note that issues that are caused by working with this package are not covered by OTRS service contracts.</p>' =>
             '',
+        '<p>The installation of packages which are not verified by the OTRS Group is not possible by default.</p>' => '<p>A instalação de pacotes que não são verificados pelo OTRS Group não é possível por omissão.</p>',
+        '<p>You can activate the installation of not verified packages in the <a href=\'$Self->{CGIHandle}?Action=AdminSystemConfiguration;Subaction=View;Setting=Package%3A%3AAllowNotVerifiedPackages\' target=\'_blank\'>System Configuration</a>.</p>' => '<p>Pode activar a instalação de pacotes não verificados na <a href=\'$Self->{CGIHandle}?Action=AdminSystemConfiguration;Subaction=View;Setting=Package%3A%3AAllowNotVerifiedPackages\' target=\'_blank\'>Configuração de Sistema</a>.</p>',
 
         # Perl Module: Kernel/System/ProcessManagement/DB/Process.pm
         'The process "%s" and all of its data has been imported successfully.' =>