+++ Kernel/Output/HTML/Notification/PackageManagerCheckNotVerifiedPackages.pm
@@ -1,0 +1,53 @@
+# --
+# Copyright (C) 2001-2018 OTRS AG, http://otrs.com/
+# --
+# This software comes with ABSOLUTELY NO WARRANTY. For details, see
+# the enclosed file COPYING for license information (AGPL). If you
+# did not receive this file, see http://www.gnu.org/licenses/agpl.txt.
+# --
+
+package Kernel::Output::HTML::Notification::PackageManagerCheckNotVerifiedPackages;
+
+use parent 'Kernel::Output::HTML::Base';
+
+use strict;
+use warnings;
+
+our @ObjectDependencies = (
+    'Kernel::Config',
+    'Kernel::Output::HTML::Layout',
+    'Kernel::System::Group',
+    'Kernel::System::SysConfig',
+);
+
+sub Run {
+    my ( $Self, %Param ) = @_;
+
+    # Check if setting is activated.
+    my $PackageAllowNotVerifiedPackages = $Kernel::OM->Get('Kernel::Config')->Get('Package::AllowNotVerifiedPackages');
+    return '' if !$PackageAllowNotVerifiedPackages;
+
+    # Check permissions.
+    my $Group = $Param{Config}->{Group} || 'admin';
+    my $HasPermission = $Kernel::OM->Get('Kernel::System::Group')->PermissionCheck(
+        UserID    => $Self->{UserID},
+        GroupName => $Group,
+        Type      => 'rw',
+    );
+
+    return '' if !$HasPermission;
+
+    my $LayoutObject = $Kernel::OM->Get('Kernel::Output::HTML::Layout');
+
+    return $LayoutObject->Notify(
+        Priority => 'Error',
+        Data     => $LayoutObject->{LanguageObject}->Translate(
+            'The installation of packages which are not verified by the OTRS Group is activated. These packages could threaten your whole system! It is recommended not to use unverified packages.'
+        ),
+        Link =>
+            $LayoutObject->{Baselink}
+            . 'Action=AdminSystemConfiguration;Subaction=View;Setting=Package%3A%3AAllowNotVerifiedPackages;',
+    );
+}
+
+1;